Fix CI across databases and static checks
