+++ tests/TestCase/Controller/Admin/ScopesControllerTest.php
@@ -3,12 +3,13 @@ declare(strict_types=1);
 
 namespace TinyAuthBackend\Test\TestCase\Controller\Admin;
 
-use Cake\ORM\TableRegistry;
 use Cake\TestSuite\IntegrationTestTrait;
 use Cake\TestSuite\TestCase;
+use TinyAuthBackend\Test\TestSuite\DatabaseTestTrait;
 
 class ScopesControllerTest extends TestCase {
 
+	use DatabaseTestTrait;
 	use IntegrationTestTrait;
 
 	protected array $fixtures = [
@@ -66,14 +67,6 @@ class ScopesControllerTest extends TestCase {
 
 		$this->assertResponseCode(302);
 		$this->assertSame(0, $this->countRows('tinyauth_scopes', ['id' => 1]));
-	}
-
-	protected function insertRow(string $table, array $data): void {
-		TableRegistry::getTableLocator()->get($table)->getConnection()->insert($table, $data);
-	}
-
-	protected function countRows(string $table, array $conditions): int {
-		return TableRegistry::getTableLocator()->get($table)->find()->where($conditions)->count();
 	}
 
 }
