+++ tests/TestCase/Policy/TinyAuthPolicyTest.php
@@ -3,18 +3,16 @@ declare(strict_types=1);
 
 namespace TinyAuthBackend\Test\TestCase\Policy;
 
-use ArrayAccess;
-use Authorization\IdentityInterface;
-use Authorization\Policy\ResultInterface;
 use Cake\Core\Configure;
-use Cake\Datasource\EntityInterface;
 use Cake\ORM\Entity;
-use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 use TestApp\Model\Entity\Article;
 use TinyAuthBackend\Policy\TinyAuthPolicy;
+use TinyAuthBackend\Test\TestSuite\DatabaseTestTrait;
 
 class TinyAuthPolicyTest extends TestCase {
+
+	use DatabaseTestTrait;
 
 	protected array $fixtures = [
 		'plugin.TinyAuthBackend.TinyAuthRoles',
@@ -85,49 +83,6 @@ class TinyAuthPolicyTest extends TestCase {
 		$result = $policy->before($identity, new Article(), 'edit');
 
 		$this->assertTrue($result);
-	}
-
-	protected function insertRow(string $table, array $data): void {
-		TableRegistry::getTableLocator()->get($table)->getConnection()->insert($table, $data);
-	}
-
-}
-
-class TestIdentity implements IdentityInterface {
-
-	public function __construct(protected EntityInterface&ArrayAccess $data) {
-	}
-
-	public function can(string $action, mixed $resource): bool {
-		return false;
-	}
-
-	public function canResult(string $action, mixed $resource): ResultInterface {
-		throw new \BadMethodCallException('Not implemented for this test double.');
-	}
-
-	public function applyScope(string $action, mixed $resource, mixed ...$optionalArgs): mixed {
-		return $resource;
-	}
-
-	public function getOriginalData(): ArrayAccess|array {
-		return $this->data;
-	}
-
-	public function offsetExists(mixed $offset): bool {
-		return $this->data->offsetExists($offset);
-	}
-
-	public function offsetGet(mixed $offset): mixed {
-		return $this->data->offsetGet($offset);
-	}
-
-	public function offsetSet(mixed $offset, mixed $value): void {
-		$this->data->offsetSet($offset, $value);
-	}
-
-	public function offsetUnset(mixed $offset): void {
-		$this->data->offsetUnset($offset);
 	}
 
 }
